requirements.txt
streamlit
pillow
PyMuPDF
google-generativeai
python-dotenv
